requirements.txt
streamlit
tensorflow==2.12.0
pillow>=7.1.0,<12
numpy
flask